chore: update file names
