+++ packages/connection-encrypter-tls/src/index.ts
@@ -18,14 +18,10 @@
  * ```
  */
 
-import { TLSSocket, type TLSSocketOptions, connect } from 'node:tls'
-import { CodeError } from '@libp2p/interface'
-import { generateCertificate, verifyPeerCertificate, itToStream, streamToIt } from './utils.js'
-import type { ComponentLogger, MultiaddrConnection, ConnectionEncrypter, SecuredConnection, PeerId, Logger } from '@libp2p/interface'
-import type { Duplex } from 'it-stream-types'
-import type { Uint8ArrayList } from 'uint8arraylist'
+import { TLS } from './tls.js'
+import type { ComponentLogger, ConnectionEncrypter } from '@libp2p/interface'
 
-const PROTOCOL = '/tls/1.0.0'
+export const PROTOCOL = '/tls/1.0.0'
 
 export interface TLSComponents {
   logger: ComponentLogger
@@ -37,94 +33,6 @@ export interface TLSInit {
    * (default: 1000)
    */
   timeout?: number
-}
-
-class TLS implements ConnectionEncrypter {
-  public protocol: string = PROTOCOL
-  private readonly log: Logger
-  private readonly timeout: number
-
-  constructor (components: TLSComponents, init: TLSInit = {}) {
-    this.log = components.logger.forComponent('libp2p:tls')
-    this.timeout = init.timeout ?? 1000
-  }
-
-  async secureInbound <Stream extends Duplex<AsyncGenerator<Uint8Array | Uint8ArrayList>> = MultiaddrConnection> (localId: PeerId, conn: Stream, remoteId?: PeerId): Promise<SecuredConnection<Stream>> {
-    return this._encrypt(localId, conn, false, remoteId)
-  }
-
-  async secureOutbound <Stream extends Duplex<AsyncGenerator<Uint8Array | Uint8ArrayList>> = MultiaddrConnection> (localId: PeerId, conn: Stream, remoteId?: PeerId): Promise<SecuredConnection<Stream>> {
-    return this._encrypt(localId, conn, true, remoteId)
-  }
-
-  /**
-   * Encrypt connection
-   */
-  async _encrypt <Stream extends Duplex<AsyncGenerator<Uint8Array | Uint8ArrayList>> = MultiaddrConnection> (localId: PeerId, conn: Stream, isServer: boolean, remoteId?: PeerId): Promise<SecuredConnection<Stream>> {
-    const opts: TLSSocketOptions = {
-      ...await generateCertificate(localId),
-      isServer,
-      // require TLS 1.3 or later
-      minVersion: 'TLSv1.3',
-      maxVersion: 'TLSv1.3',
-      // accept self-signed certificates
-      rejectUnauthorized: false
-    }
-
-    let socket: TLSSocket
-
-    if (isServer) {
-      // @ts-expect-error docs say this is fine?
-      socket = new TLSSocket(itToStream(conn), {
-        ...opts,
-        // require clients to send certificates
-        requestCert: true
-      })
-    } else {
-      socket = connect({
-        socket: itToStream(conn),
-        ...opts
-      })
-    }
-
-    return new Promise((resolve, reject) => {
-      const abortTimeout = setTimeout(() => {
-        socket.destroy(new CodeError('Handshake timeout', 'ERR_HANDSHAKE_TIMEOUT'))
-      }, this.timeout)
-
-      const verifyRemote = (): void => {
-        const remote = socket.getPeerCertificate()
-
-        verifyPeerCertificate(remote.raw, remoteId, this.log)
-          .then(remotePeer => {
-            this.log('remote certificate ok, remote peer %p', remotePeer)
-
-            resolve({
-              remotePeer,
-              conn: {
-                ...conn,
-                ...streamToIt(socket)
-              }
-            })
-          })
-          .catch(err => {
-            reject(err)
-          })
-          .finally(() => {
-            clearTimeout(abortTimeout)
-          })
-      }
-
-      socket.on('error', err => {
-        reject(err)
-        clearTimeout(abortTimeout)
-      })
-      socket.on('secure', (evt) => {
-        this.log('verifying remote certificate')
-        verifyRemote()
-      })
-    })
-  }
 }
 
 export function tls (init?: TLSInit): (components: TLSComponents) => ConnectionEncrypter {
